apply mapping to master branch
